+++ packages/chonky/src/types/icons.types.ts
@@ -18,6 +18,7 @@ export enum ChonkyIconName {
   paste = 'paste',
   share = 'share',
   search = 'search',
+  filter = 'filter',
   selectAllFiles = 'selectAllFiles',
   clearSelection = 'clearSelection',