+++ packages/chonky/src/util/file-browser-handle.ts
@@ -2,7 +2,7 @@ import React, { useImperativeHandle } from 'react';
 import { useChonkyDispatch, useChonkyReduxStore } from '../redux/store';
 
 import { reduxActions } from '../redux/reducers';
-import { selectSelectionMap } from '../redux/selectors';
+import { selectSelectionMap, selectors } from '../redux/selectors';
 import { thunkRequestFileAction } from '../redux/thunks/dispatchers.thunks';
 import { FileAction } from '../types/action.types';
 import { FileBrowserHandle } from '../types/file-browser.types';
@@ -23,6 +23,13 @@ export const useFileBrowserHandle = (ref: React.Ref<FileBrowserHandle>) => {
       setFileSelection(selection, reset = true): void {
         const fileIds = Array.from(selection);
         dispatch(reduxActions.selectFiles({ fileIds, reset }));
+      },
+      revealFile(id): void {
+        const displayFileIds = selectors.getDisplayFileIds(store.getState());
+        if (!displayFileIds.includes(id)) return;
+
+        dispatch(reduxActions.selectFiles({ fileIds: [id], reset: true }));
+        dispatch(reduxActions.revealFile(id));
       },
       async requestFileAction<Action extends FileAction>(
         action: Action,